Major revision for SPS and other smaller changes
Changed SPS to use only the 6 defined levels. Created a new component to
help with the smaller SPS values. Updated the clock system to get better
clock divisions for the ADC.

When changing the clocks, had to update the UART module. Now runs on new
baud rate.

Updated frame prefix to include SPS.